refactor: PerformanceCheckResponseDto 패키지 수정
